+++ source/AskFi.Runtime/Scraper.cs
@@ -1,5 +1,6 @@
 using AskFi.Runtime.Messages;
 using AskFi.Runtime.Modules.Observation;
+using AskFi.Runtime.Modules.Output;
 using AskFi.Runtime.Modules.Perspective;
 using AskFi.Runtime.Platform;
