+++ source/AskFi.Runtime/Sequencer.cs
@@ -1,4 +1,6 @@
 using AskFi.Runtime.Messages;
+using AskFi.Runtime.Modules.Input;
+using AskFi.Runtime.Modules.Output;
 using AskFi.Runtime.Modules.Perspective;
 using AskFi.Runtime.Platform;
